main<-Develop

+++ src/main/java/com/hrth/ustock/service/PortfolioService.java
@@ -3,27 +3,30 @@ package com.hrth.ustock.service;
 import com.hrth.ustock.dto.holding.HoldingEmbedDto;
 import com.hrth.ustock.dto.holding.HoldingRequestDto;
 import com.hrth.ustock.dto.portfolio.PortfolioListDto;
-import com.hrth.ustock.dto.portfolio.PortfolioRequestDTO;
+import com.hrth.ustock.dto.portfolio.PortfolioRequestDto;
 import com.hrth.ustock.dto.portfolio.PortfolioResponseDto;
 import com.hrth.ustock.dto.portfolio.PortfolioUpdateDto;
 import com.hrth.ustock.entity.User;
 import com.hrth.ustock.entity.portfolio.Holding;
 import com.hrth.ustock.entity.portfolio.Portfolio;
 import com.hrth.ustock.entity.portfolio.Stock;
-import com.hrth.ustock.exception.HoldingNotFoundException;
-import com.hrth.ustock.exception.PortfolioNotFoundException;
-import com.hrth.ustock.exception.StockNotFoundException;
-import com.hrth.ustock.exception.UserNotFoundException;
+import com.hrth.ustock.exception.*;
 import com.hrth.ustock.repository.HoldingRepository;
 import com.hrth.ustock.repository.PortfolioRepository;
 import com.hrth.ustock.repository.StockRepository;
 import com.hrth.ustock.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Map;
 
+import static com.hrth.ustock.service.StockServiceConst.*;
+
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class PortfolioService {
@@ -31,7 +34,8 @@ public class PortfolioService {
     private final UserRepository userRepository;
     private final HoldingRepository holdingRepository;
     private final StockRepository stockRepository;
-    private static final int tempCurrent = 100000;
+    private final RedisTemplate<String, String> redisTemplate;
+    private final StockService stockService;
 
     @Transactional(readOnly = true)
     public PortfolioListDto getPortfolioList(Long userId) {
@@ -43,14 +47,13 @@ public class PortfolioService {
             throw new PortfolioNotFoundException();
 
         list.forEach(p -> {
-            // TODO: 현재가 반영
-            refreshPortfolio(p, tempCurrent);
+            refreshPortfolio(p);
             portfolioListDto.setBudget(portfolioListDto.getBudget() + p.getBudget());
             portfolioListDto.setPrincipal(portfolioListDto.getPrincipal() + p.getPrincipal());
             portfolioListDto.setRet(portfolioListDto.getRet() + p.getRet());
             portfolioListDto.getList().add(p.toPortfolioDto());
         });
-        double ror = (portfolioListDto.getPrincipal()==0L) ? 0.0 : (double) portfolioListDto.getRet() / portfolioListDto.getPrincipal() * 100;
+        double ror = portfolioListDto.getPrincipal() == 0L ? 0.0 : (double) portfolioListDto.getRet() / portfolioListDto.getPrincipal() * 100;
         portfolioListDto.setRor(ror);
 
         return portfolioListDto;
@@ -63,108 +66,97 @@ public class PortfolioService {
         Portfolio portfolio = portfolioRepository.findById(pfId).orElseThrow(PortfolioNotFoundException::new);
         List<Holding> holdings = portfolio.getHoldings();
 
-        // TODO: 현재가 추가 필요
-        refreshPortfolio(portfolio, tempCurrent);
         if (holdings == null || holdings.isEmpty()) {
             portfolioResponseDto.setName(portfolio.getName());
             return portfolioResponseDto;
         }
+        refreshPortfolio(portfolio);
 
-        // 현재가 반영전까진 현재 정보 기준으로 ror 계산
-        // 현재가 반영후에는 현재가로 ret 갱신, ror 계산 후 save
-        holdings.forEach(h -> {
+        // 현재가로 ret 갱신, ror 계산 후 save
+        for(Holding h : holdings) {
+            Map<String, String> redisMap =  stockService.getCurrentChangeChangeRate(h.getStock().getCode());
+            if(redisMap == null) {
+                log.info("no current for stock: {}", h.getStock().getCode());
+                continue;
+            }
+            int quantity = h.getQuantity();
+            int average = h.getAverage();
+            int current = Integer.parseInt(redisMap.get(REDIS_CURRENT_KEY));
+            long ret = (long) quantity * current - (long) quantity * average;
+
+
             Stock stock = h.getStock();
-            long ret = (long) h.getQuantity() * tempCurrent - (long) h.getQuantity() * h.getAverage();
             HoldingEmbedDto holdingEmbedDto = HoldingEmbedDto.builder()
                     .code(stock.getCode())
                     .name(stock.getName())
-                    .quantity(h.getQuantity())
-                    .average(h.getAverage())
-                    .ror((h.getQuantity()*h.getAverage()==0) ? 0.0 : (double)(ret / ((long) h.getQuantity() * h.getAverage())) * 100)
+                    .logo(stock.getLogo())
+                    .quantity(quantity)
+                    .average(average)
+                    .ror((quantity * average == 0) ? 0.0 : (double) ret / ((long) quantity * average) * 100)
                     .build();
             portfolioResponseDto.getStocks().add(holdingEmbedDto);
-        });
+        }
 
         portfolioResponseDto.setName(portfolio.getName());
         portfolioResponseDto.setBudget(portfolio.getBudget());
         portfolioResponseDto.setPrincipal(portfolio.getPrincipal());
         portfolioResponseDto.setRet(portfolio.getRet());
-        portfolioResponseDto.setRor((portfolio.getPrincipal()==0L) ? 0.0 : (double) portfolio.getRet() / portfolio.getPrincipal() * 100);
+        portfolioResponseDto.setRor((portfolio.getPrincipal() == 0L) ? 0.0 : (double) portfolio.getRet() / portfolio.getPrincipal() * 100);
         return portfolioResponseDto;
     }
 
     @Transactional
-    public void addPortfolio(PortfolioRequestDTO portfolioRequestDTO, Long userId) {
+    public void addPortfolio(PortfolioRequestDto portfolioRequestDto, Long userId) {
 
         User user = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
         Portfolio portfolio = Portfolio.builder()
-                .name(portfolioRequestDTO.getName())
+                .name(portfolioRequestDto.getName())
                 .user(user)
                 .budget(0L)
                 .principal(0L)
                 .ret(0L)
                 .build();
+
         portfolioRepository.save(portfolio);
+    }
+
+    @Transactional
+    public void buyStock(Long pfId, String code, HoldingRequestDto holdingRequestDto) {
+        Holding testHolding = holdingRepository.findHoldingByPortfolioIdAndStockCode(pfId, code).orElse(null);
+
+        if (testHolding != null) {
+            additionalBuyStock(pfId, code, holdingRequestDto);
+            return;
+        }
+
+        Portfolio portfolio = portfolioRepository.findById(pfId).orElseThrow(PortfolioNotFoundException::new);
+        Stock stock = stockRepository.findByCode(code).orElseThrow(StockNotFoundException::new);
+        Holding holding = Holding.builder()
+                .portfolio(portfolio)
+                .stock(stock)
+                .average(holdingRequestDto.getPrice())
+                .quantity(holdingRequestDto.getQuantity())
+                .build();
+
+        holdingRepository.save(holding);
     }
 
     // 10. 개별 종목 추가 매수
     @Transactional
-    public void buyStock(Long pfId, String code, HoldingRequestDto holdingRequestDto) {
+    public void additionalBuyStock(Long pfId, String code, HoldingRequestDto holdingRequestDto) {
 
-        Portfolio portfolio = portfolioRepository.findById(pfId).orElseThrow(PortfolioNotFoundException::new);
-        List<Holding> holdings = portfolio.getHoldings();
+        Holding target = holdingRepository.findHoldingByPortfolioIdAndStockCode(pfId, code).orElseThrow(HoldingNotFoundException::new);
 
-        Holding target = null;
-        // 보유 종목 중에서 탐색 후 매수, 없으면 새로운 보유 종목 추가
-        for (Holding h : holdings) {
-            if (h.getStock().getCode().equals(code)) {
-                target = h;
-            }
-        }
-        if (target != null) {
-            target.additionalBuyHolding(holdingRequestDto.getQuantity(), holdingRequestDto.getPrice());
-            holdingRepository.save(target);
-        } else {
-            Stock stock = stockRepository.findByCode(code).orElseThrow(StockNotFoundException::new);
-            Holding newHolding = Holding.builder()
-                    .stock(stock)
-                    .portfolio(portfolio)
-                    .quantity(holdingRequestDto.getQuantity())
-                    .average(holdingRequestDto.getPrice())
-                    .build();
-            holdingRepository.save(newHolding);
-        }
+        target.additionalBuyHolding(holdingRequestDto.getQuantity(), holdingRequestDto.getPrice());
     }
 
     // 11. 개별 종목 수정
     @Transactional
     public void editHolding(Long pfId, String code, HoldingRequestDto holdingRequestDto) {
 
-        Portfolio portfolio = portfolioRepository.findById(pfId).orElseThrow(PortfolioNotFoundException::new);
-        List<Holding> holdings = portfolio.getHoldings();
+        Holding target = holdingRepository.findHoldingByPortfolioIdAndStockCode(pfId, code).orElseThrow(HoldingNotFoundException::new);
 
-        Holding target = null;
-        // 보유 종목 중에서 탐색 후 있으면 삭제, 없으면 throw exception
-        for (Holding h : holdings) {
-            if (h.getStock().getCode().equals(code)) {
-                target = h;
-            }
-        }
-        if (target != null) {
-            // input: quantity, price = 0일시 해당 보유종목 삭제
-            if (holdingRequestDto.getQuantity() == 0 || holdingRequestDto.getPrice() == 0) {
-                holdingRepository.delete(target);
-                portfolio.getHoldings().remove(target);
-            } else {
-                target.updateHolding(holdingRequestDto.getQuantity(), holdingRequestDto.getPrice());
-                holdingRepository.save(target);
-            }
-        } else {
-            throw new HoldingNotFoundException();
-        }
-
-        // TODO: 현재가 추가
-        refreshPortfolio(portfolio, tempCurrent);
+        target.updateHolding(holdingRequestDto.getQuantity(), holdingRequestDto.getPrice());
     }
 
     // 12. 개별 종목 삭제
@@ -172,21 +164,11 @@ public class PortfolioService {
     public void deleteHolding(Long pfId, String code) {
 
         Portfolio portfolio = portfolioRepository.findById(pfId).orElseThrow(PortfolioNotFoundException::new);
-        List<Holding> holdings = portfolio.getHoldings();
+      
+        Holding target = holdingRepository.findHoldingByPortfolioIdAndStockCode(pfId, code).orElseThrow(HoldingNotFoundException::new);
 
-        Holding target = null;
-        // 보유 종목 중에서 탐색 후 있으면 삭제, 없으면 throw exception
-        for (Holding h : holdings) {
-            if (h.getStock().getCode().equals(code)) {
-                target = h;
-            }
-        }
-        if (target != null) {
-            holdingRepository.delete(target);
-            portfolio.getHoldings().remove(target);
-        } else {
-            throw new HoldingNotFoundException();
-        }
+        holdingRepository.delete(target);
+        portfolio.getHoldings().remove(target);
     }
 
     // 13. 포트폴리오 삭제
@@ -197,34 +179,35 @@ public class PortfolioService {
         List<Holding> list = portfolio.getHoldings();
 
         // 보유 종목 없을시 바로 삭제, 아니면 보유종목 전부 삭제
-        if (list == null || list.isEmpty()) {
-            portfolioRepository.delete(portfolio);
-        } else {
+        if (list != null && !list.isEmpty()) {
             holdingRepository.deleteAll(list);
-            portfolioRepository.delete(portfolio);
         }
+
+        portfolioRepository.delete(portfolio);
     }
 
     @Transactional
-    protected void refreshPortfolio(Portfolio portfolio, int current) {
-
-        // TODO: 현재가 변경 - holdings의 stock별 현재가
+    protected void refreshPortfolio(Portfolio portfolio) {
         PortfolioUpdateDto portfolioUpdateDto = new PortfolioUpdateDto();
         List<Holding> list = portfolio.getHoldings();
 
-        if (list == null || list.isEmpty()) {
-            portfolioUpdateDto.setBudget(portfolio.getBudget());
-            portfolio.updatePortfolio(portfolioUpdateDto);
-            return;
-        }
+        if (list == null || list.isEmpty()) return;
 
         // budget = 원금+수익, principal = 갯수+평단가, ret = 갯수+현재가
-        list.forEach(h -> {
+        for(Holding h : list) {
+            Map<String, String> redisMap = stockService.getCurrentChangeChangeRate(h.getStock().getCode());
+            if(redisMap == null) {
+                log.info("no current for stock: {}", h.getStock().getCode());
+                continue;
+            }
+            int current = Integer.parseInt(redisMap.get(REDIS_CURRENT_KEY));
+
             long before = (long) h.getQuantity() * h.getAverage();
             long after = (long) h.getQuantity() * current;
             portfolioUpdateDto.setPrincipal(portfolioUpdateDto.getPrincipal() + before);
             portfolioUpdateDto.setRet(portfolioUpdateDto.getRet() + after - before);
-        });
+        }
+
         portfolioUpdateDto.setBudget(portfolioUpdateDto.getPrincipal() + portfolioUpdateDto.getRet());
         portfolio.updatePortfolio(portfolioUpdateDto);
         portfolioRepository.save(portfolio);
